requirements.txt
click==8.1.3
Flask==2.2.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mysqlclient==2.1.1
packaging==23.0
PyJWT==2.6.0
six==1.16.0
SQLAlchemy==2.0.4
typing_extensions==4.5.0
Werkzeug==2.2.3
zipp==3.15.0
